Passing unit tests

+++ data/example.py
@@ -1,25 +1,22 @@
-import os
-from errors import Missing
-
-os.environ["EXAMPLE_DB"] = ":memory:"
-
-from data.init_sqlite import curs
+from .init_sqlite import curs, IntegrityError
 from model.example import Example
+from errors import Missing, Duplicate
 
 curs.execute(
     """create table if not exists example(
-            name text primary key,
-            country text,
-            description text)"""
+                name primary key,
+                country text,
+                description text)"""
 )
 
 
 def row_to_model(row: tuple) -> Example:
-    return Example(name=row[0], country=row[1], description=row[2])
+    name, country, description = row
+    return Example(name=name, country=country, description=description)
 
 
 def model_to_dict(example: Example) -> dict:
-    return example.model_dump() if example else None
+    return example.dict()
 
 
 def get_one(name: str) -> Example:
@@ -28,7 +25,7 @@ def get_one(name: str) -> Example:
     curs.execute(qry, params)
     row = curs.fetchone()
     if row:
-        return row_to_model(row=row)
+        return row_to_model(row)
     else:
         raise Missing(msg=f"Example {name} not found")
 
@@ -40,28 +37,40 @@ def get_all() -> list[Example]:
 
 
 def create(example: Example) -> Example:
-    qry = """insert into example (name, country, description)
-    values (:name, :country, :description)"""
+    if not example:
+        return None
+    qry = """insert into example
+             (name, country, description) values
+             (:name, :country, :description)"""
     params = model_to_dict(example)
-    _ = curs.execute(qry, params)
+    try:
+        curs.execute(qry, params)
+    except IntegrityError:
+        raise Duplicate(msg=f"Example {example.name} already exists")
     return get_one(example.name)
 
 
 def modify(name: str, example: Example) -> Example:
+    if not (name and example):
+        return None
     qry = """update example
-    set country=:country,
-    name=:name,
-    description=:description
-    where name=:name_orig"""
+             set name=:name, country=:country,
+                 description=:description
+             where name=:orig_name"""
     params = model_to_dict(example)
-    params["name_orig"] = example.name
-    _ = curs.execute(qry, params)
-    explorer2 = get_one(example.name)
-    return explorer2
+    params["orig_name"] = name
+    curs.execute(qry, params)
+    if curs.rowcount == 1:
+        return get_one(example.name)
+    else:
+        raise Missing(msg=f"Example {name} not found")
 
 
-def delete(example: Example) -> bool:
+def delete(name: str):
+    if not name:
+        return False
     qry = "delete from example where name = :name"
-    params = {"name": example.name}
-    res = curs.execute(qry, params)
-    return bool(res)
+    params = {"name": name}
+    curs.execute(qry, params)
+    if curs.rowcount != 1:
+        raise Missing(msg=f"Example {name} not found")
